pyproject.toml
[project]
name = "devices-db"
version = "0.1.0"
description = "Flask web application for devices database with OAuth2 authentication"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.0.0",
    "pandas>=2.0.0",
    "openpyxl>=3.1.0",
    "authlib>=1.2.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.0",
    "msal>=1.33.0",
]
